+++ apps/web/src/app/partner/_shared/api/get-profile.ts
@@ -1,7 +1,7 @@
 import "@/shared/api/client";
 import { getProfile as getProfileRequest } from "@/shared/api/generated/sdk.gen";
-import { adjusterProfileSchema } from "../_model/profile.schema";
-import type { AdjusterProfile } from "../_model/types";
+import { adjusterProfileSchema } from "../model/adjuster-profile.schema";
+import type { AdjusterProfile } from "../model/adjuster-profile.schema";
 
 export async function getProfile(): Promise<AdjusterProfile> {
   const { data } = await getProfileRequest({ throwOnError: true });
