--- apps/web/src/app/partner/profile/edit/_api/get-profile.ts
@@ -1,9 +1,0 @@
-import "@/shared/api/client";
-import { getProfile as getProfileRequest } from "@/shared/api/generated/sdk.gen";
-import { adjusterProfileSchema } from "../_model/adjuster-profile.schema";
-import type { AdjusterProfile } from "../_model/types";
-
-export async function getProfile(): Promise<AdjusterProfile> {
-  const { data } = await getProfileRequest({ throwOnError: true });
-  return adjusterProfileSchema.parse(data);
-}
