+++ apps/web/src/app/partner/profile/edit/_api/update-profile.ts
@@ -1,7 +1,8 @@
 import "@/shared/api/client";
 import { updateProfile as updateProfileRequest } from "@/shared/api/generated/sdk.gen";
-import { adjusterProfileSchema } from "../_model/adjuster-profile.schema";
-import type { AdjusterProfile, UpdateProfileBody } from "../_model/types";
+import { adjusterProfileSchema } from "@/app/partner/_shared/model/adjuster-profile.schema";
+import type { AdjusterProfile } from "@/app/partner/_shared/model/adjuster-profile.schema";
+import type { UpdateProfileBody } from "../_model/types";
 
 export async function updateProfile(body: UpdateProfileBody): Promise<AdjusterProfile> {
   const { data } = await updateProfileRequest({
