--- apps/web/src/app/partner/profile/edit/_api/use-profile.ts
@@ -1,15 +1,0 @@
-"use client";
-
-import { useSuspenseQuery } from "@tanstack/react-query";
-import { adjusterKeys } from "@/shared/api/query-keys";
-import { GC_TIME_DEFAULT, STALE_TIME_AUTH } from "@/shared/api/query-constants";
-import { getProfile } from "./get-profile";
-
-export function useProfile() {
-  return useSuspenseQuery({
-    queryKey: adjusterKeys.meProfile().queryKey,
-    queryFn: getProfile,
-    staleTime: STALE_TIME_AUTH,
-    gcTime: GC_TIME_DEFAULT,
-  });
-}
